requirements.txt
Brotli==1.1.0
Django==4.2.13
django-cleanup==8.1.0
django-filter==24.2
django-quill-editor==0.1.40
django-simple-history==3.7.0
mssql-django==1.5
pillow==10.3.0
python-dotenv==1.0.1
setuptools==68.2.2
waitress==3.0.0
whitenoise==6.6.0